aim2_metrics/requirements.txt
appdirs==1.4.4
arrow==0.17.0
attrs==20.1.0
black==20.8b1
cfgv==3.2.0
click==7.1.2
ConfigArgParse==1.2.3
devtools==0.6
distlib==0.3.1
filelock==3.0.12
flake8==3.8.3
identify==1.4.30
importlib-metadata==1.7.0
iniconfig==1.0.1
isort==5.4.2
loguru==0.5.3
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.782
mypy-extensions==0.4.3
nodeenv==1.5.0
numpy==1.19.1
opencv-python==4.4.0.42
packaging==20.4
pandas==1.2.1
pathspec==0.8.0
Pillow==7.2.0
pluggy==0.13.1
pre-commit==2.7.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==3.0.0a2
pytest==6.0.1
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
virtualenv==20.0.31
zipp==3.1.0
